Added support for 64-bit virtual addresses.

+++ src/include/base/Types.h
@@ -36,7 +36,7 @@ typedef long ssize_t;
 #endif
 
 typedef unsigned long word_t;
-typedef uint32_t label_t;
+typedef uint64_t label_t;
 typedef uint64_t capsel_t;
 typedef int fd_t;
 typedef uint64_t cycles_t;
